v1.2try-6 现在支持<World>变量 支持以op身份执行指令 $c3 支持延时执行后面的内容 $d数字 修复任何人可删除菜单的bug

+++ src/main/java/com/trally/GreatMenu.java
@@ -100,21 +100,23 @@ public class GreatMenu extends JavaPlugin {
                 }
 
                 if (args[0].equals("remove")) {
-                    if (menus.containsKey(args[1])) {
-                        File tmp = new File(menuFolder, args[1] + ".yml");
-                        try {
-                            if (tmp.delete()) {
-                                p.sendMessage("§a删除成功");
-                                reLoadMenus();
-                            } else {
+                    if(p.isOp()){
+                        if (menus.containsKey(args[1])) {
+                            File tmp = new File(menuFolder, args[1] + ".yml");
+                            try {
+                                if (tmp.delete()) {
+                                    p.sendMessage("§a删除成功");
+                                    reLoadMenus();
+                                } else {
+                                    p.sendMessage("§a可能由于文件被打开等原因，删除失败");
+                                }
+                            } catch (Exception e) {
                                 p.sendMessage("§a可能由于文件被打开等原因，删除失败");
+                                throw new RuntimeException(e);
                             }
-                        } catch (Exception e) {
-                            p.sendMessage("§a可能由于文件被打开等原因，删除失败");
-                            throw new RuntimeException(e);
+                        } else {
+                            p.sendMessage("§4没有此菜单");
                         }
-                    } else {
-                        p.sendMessage("§4没有此菜单");
                     }
                 }
